Author: R. Mitchell Changes: Numerous. Added all necessary stages to create a real-time graphing node for the multi-modal encodings. The MMCX encodings for both light and wind are now shunted out to a python graphing node which plots both as well as the CL representation (and has a "true" CL representation too, currently disabled).
I tested this on the robot with a live wind cue and fake light cue.
I haven't yet tested it with two real cues which I think is kinda
necessary in order to evaluate the output. This can wait until next
week. Pushing largely so I don't lose what feels like an immense
amount of progress in the last week.

Testing:
Graphed output with both dummy and half-real cues. Seemed to make
sense but I need more information to learn how to interpret this.
Right now it doesn't look like an average but there's now so many
factors at play I need to sit down and figure out what is fully
happening to get to the plot. Super happy with this though.