Author; R. Mitchell Changes: Added cue calibration to dummy cue nodes in order to aid testing. Fixed some broken elements of the calibration routine.
Testing:
Tested calibration in simulation and it seems to work as intended.
Still needs tested on the actual robot which should happen later today.
Main point of difference will likely be that the offsets are applied in
the cue detection nodes so need to watch for errors there.

Notes:
There isn't really a way to clear calibration without doing it manually
by killing the detection nodes and then wiping the parameters (or restarting
roscore). I tried to add this in cue_calibration but it doesn't work; I think
it gets a message from the cue which contains the "calibrated" azimuth
then tries to calibrate that rather than the world azimuth. Thus the
re-calibration routine, for the moment has to be:

stop detection nodes
clear calibration (by running rosparam delete)
restart detection nodes

Feels ridiculously cumbersome but I could probably write a shell script
to do it.