Added a pol cue detection node.
I've re-added the pol cue detector but for the moment the angular
decoding doesn't seem to be working and I'm not sure why. I tested
this by playing back some of the sardinia rosbags and watching the
output.

The detector does not currently create and publish a Cue message but
adding this should be straightforward enough.