Added an isnan check to the TL2 neurons.
The polarisation sensor can sometimes return NaN due to poor sky
readings. If this is passed as angular input to the CX then NaNs
persist in the CX due to the memory circuit and it becomes
impossible to clear. I've added a check to the TL2 inputs to check
if the angle passed is NaN, if it is, the code should set all TL2
neurons to 0.

This is currently untested.