Made a little headway on the erm visualisation
Got the require message types in place and wrote the publishing code.
It doesn't work in this commit as there is an issue with packing
the 2D mappings into messages but I think there's a solution
where you explicitly cast everything to float.
I just don't have the mental energy to figure out the 'pythonic'
way of doing this at the moment.