Vis tools plus bug fixes
Built the vis script for the network and noticed an issue with the
neural rates. There were a combination of factors but this has
now been fixed.

There is still a problem where the final position of cues on init
has to match the position of the cues which then take over after
init and I have yet to find a solution to this. However,
as in the older simulations, spinning the robot around at high speeds
seems to work for getting a bump back after its been lost.