Mixed changeset: fixed bb_computation CMake and VM modifications
Updated the CMakeLists file for bb_computation to exclude
some nodes which it isn't worth compiling. Specifically those
which pertain to the old MMCX model. I'm wondering if it's even
worth including this code.

I also modified the test_vmcx_subscriber so that it can operate easily
from the command line which could make for a nice demo. Unfortunately
the homing direction works but the distance doesn't seem to get integrated
on the way back. Not sure why as it seems to work perfectly on the way out.

It would be good to add a launcher file so that this can be run
easily as a demo.