Found and fixed the problem with the cue manager node.
There was an issue where this node was generating segmentation
faults. It wasn't a problem with the node but I've added an
error check and an informative error message for this case.

The cue manager expected only wind and intensity cue types and
was not prepared to deal with 'dummy_cue's. The check will
catch the case where the cue type was not recognised and
give the user a warning about cue types and what the likely
problem is.