Completed documentation for bb_computation
I think the docs for bb_computation are now complete with two
(fine) exceptions.

The first is the MMCX. I've included enough so that someone can
see what it's meant to do but included a bunch of warnings
not to use it because it's outdated.

The second is that I haven't documented the C++ ERM implementation
because it's so far from being useful that there's no point.
I've included a warning stating that the file is not yet ready.