In the middle of writing control code for modality transfer
Updated the ERM class to have a steering function which (for the
moment) computes the signed angle between the E-PG angle and the
goal angle.

Also started writing the control routine for the experiments.
Added keyboard support (which has also added a dependency) which allows
an operator to command the robot (though this needs to be better
implemented, stopping needs to be listened for everywhere).
I've also written a working dance routine so can write menotaxis
tomorrow.