Bio-inspired steering circuit seems to be working
The bio-inspired (but anatomically inaccurate) steering circuit
appears to work without too many headaches. Tuning the steering to
work with its output was a little awkward.

Note: I need to check that the R input actually modifies the EPGs
during behaviour as I have a suspicion that the PENs are doing all
the work (and the accuracy of the odometry is essentially masking
a flawed model).