Added PID control and more fine-grained steering control for menotaxis
PID control means we can be more selective with the error threshold
for forward motion. The controller is definitely sensitve to changes
in the integral term which cause big sweeps in the robot's path.
I'm guessing this is because the PFL3L/R comparison isn't an 'error'
in the sense which is usually meant with PID.