Stable working version
This version could run the experiment but there's still this ongoing
problem with EPG rotation and real-world rotation not matching.

This version seems the best so far in terms of repeatable orientation
but it's hard to really test without a proper wind set up.

The bump can die in this version but it's not a serious problem in
real testing, it's more of a simulation issue.